requirements.txt
Flask
tensorflow
Pillow
requests